Updated to have actionPerformed in controller
